custom jobs start in project path by default
