--- src/model/cloud-runner/services/cloud-runner-override-sources/gcp-secret-manager.yaml
@@ -1,2 +1,0 @@
-name: gcp-secret-manager
-read-secret: gcloud secrets versions access 1 --secret=\"{0}\"
